Venmo Desktop - Consumer Feedback collection (#1747)
* Add onSubmitFeedback to the QRCodeComponent props.

* Update the container height and fix flow issue.

* Fix lint issue.

* Explicitly set the onSubmitFeedback to not be required.

* Fix lint issue.

* Explicitly set the onSubmitFeedback to not be required.

* Meve the close button to be handled by the prerenderTemplate.

* Fix lint issues.

* Fix flow issue.

* send env, sessionID, locale, clientID, sdkCorrelationID, and buyerCountry as xprops to child component.

* increase bundle size limit to 75kb.

* remove unecessary close button style and remove function.

* refactor

* revert size check max

* revert size check max

* refactor

* revert size check max

* remove onSubmitFeedback

* add buttonSessionID to xprops for logging

* add buttonSessionID

* add orderID as prop

* add required

* build spinner

* refactor

* refactor

* uncomment destroy

* fix order of frames
