+++ src/zoid/qr-code/component.jsx
@@ -1,24 +1,13 @@
 /* @flow */
 /** @jsx node */
 
-import { inlineMemoize, destroyElement, type EventEmitterType } from 'belter/src';
-import { create, EVENT, type ZoidComponent } from 'zoid/src';
-import { node, dom, type ChildType } from 'jsx-pragmatic/src';
-import type { ZalgoPromise } from 'zalgo-promise/src';
-import { getLogger, getPayPalDomainRegex, getPayPalDomain, getCSPNonce, getSDKMeta, getDebug } from '@paypal/sdk-client/src';
-import { SpinnerPage } from '@paypal/common-components/src';
+import { inlineMemoize, uniqueID } from 'belter/src';
+import { create, type ZoidComponent } from 'zoid/src';
+import { getLogger, getPayPalDomainRegex, getPayPalDomain, getCSPNonce, getSDKMeta, getDebug, getEnv, getSessionID, getLocale, getClientID, getCorrelationID, getBuyerCountry } from '@paypal/sdk-client/src';
 
-const CLASS = {
-    VISIBLE:         'visible',
-    INVISIBLE:       'invisible',
-    COMPONENT_FRAME: 'component-frame',
-    PRERENDER_FRAME: 'prerender-frame'
-};
-
-type QRCodeProps = {|
-    qrPath : string,
-    cspNonce : ?string
-|};
+import { containerTemplate } from './container';
+import { prerenderTemplate } from './prerender';
+import { type QRCodeProps } from './types';
 
 export type QRCodeComponent = ZoidComponent<QRCodeProps>;
 
@@ -32,34 +21,10 @@ export function getQRCodeComponent() : QRCodeComponent {
                 width:  '100%',
                 height: '100%'
             },
-            logger:            getLogger(),
-            prerenderTemplate: ({ doc, props }) => {
-                return (
-                    <SpinnerPage
-                        nonce={ props.cspNonce }
-                    />
-                ).render(dom({ doc }));
-            },
+            logger:     getLogger(),
+            prerenderTemplate,
 
-            containerTemplate: ({ close, frame, prerenderFrame, props, doc, uid, event }) => {
-                if (!frame || !prerenderFrame) {
-                    return;
-                }
-
-                const { cspNonce } = props;
-
-                return (
-                    <QRCodeContainer
-                        uid={ uid }
-                        cspNonce={ cspNonce }
-                        close={ close }
-                        event={ event }
-                        frame={ frame }
-                        prerenderFrame={ prerenderFrame }
-                    />
-                    
-                ).render(dom({ doc }));
-            },
+            containerTemplate,
             autoResize: {
                 width:  true,
                 height: true
@@ -88,7 +53,8 @@ export function getQRCodeComponent() : QRCodeComponent {
                 debug: {
                     type:       'boolean',
                     value:      getDebug,
-                    queryParam: true
+                    queryParam: true,
+                    required:   false
                 },
                 state: {
                     type:       'string',
@@ -105,127 +71,55 @@ export function getQRCodeComponent() : QRCodeComponent {
                     queryParam:  true,
                     sendToChild: false,
                     value:       getSDKMeta
+                },
+                env: {
+                    type:        'string',
+                    queryParam:  true,
+                    required:    false,
+                    value:       getEnv
+                },
+                sessionID: {
+                    type:        'string',
+                    queryParam:  true,
+                    required:    false,
+                    value:       getSessionID
+                },
+                locale: {
+                    type:        'object',
+                    queryParam:  true,
+                    required:    false,
+                    value:       getLocale
+                },
+                clientID: {
+                    type:         'string',
+                    queryParam:   true,
+                    required:     false,
+                    value:        getClientID
+                },
+                orderID: {
+                    type:       'string',
+                    queryParam: true,
+                    required:   false
+                },
+                sdkCorrelationID: {
+                    type:       'string',
+                    required:   false,
+                    value:      getCorrelationID,
+                    queryParam: true
+                },
+                buyerCountry: {
+                    type:       'string',
+                    queryParam: true,
+                    required:   false,
+                    value:      getBuyerCountry
+                },
+                buttonSessionID: {
+                    type:       'string',
+                    queryParam: true,
+                    required:   false,
+                    value:      uniqueID
                 }
             }
         });
     });
-}
-
-
-export function QRCodeContainer({
-    uid,
-    frame,
-    prerenderFrame,
-    event,
-    cspNonce,
-    close
-} : {|
-    uid : string,
-    frame : ?HTMLIFrameElement,
-    prerenderFrame : ?HTMLIFrameElement,
-    event : EventEmitterType,
-    cspNonce? : ?string,
-    close : () => ZalgoPromise<void>
-|}) : ?ChildType {
-    if (!frame || !prerenderFrame) {
-        throw new Error(`Expected frame and prerenderframe`);
-    }
-
-    frame.classList.add(CLASS.COMPONENT_FRAME);
-    prerenderFrame.classList.add(CLASS.PRERENDER_FRAME);
-
-    frame.classList.add(CLASS.INVISIBLE);
-    prerenderFrame.classList.add(CLASS.VISIBLE);
-
-    event.on(EVENT.RENDERED, () => {
-        prerenderFrame.classList.remove(CLASS.VISIBLE);
-        prerenderFrame.classList.add(CLASS.INVISIBLE);
-
-        frame.classList.remove(CLASS.INVISIBLE);
-        frame.classList.add(CLASS.VISIBLE);
-
-        setTimeout(() => {
-            destroyElement(prerenderFrame);
-        }, 1000);
-    });
-
-    return (
-        <div id={ uid }>
-            <style
-                nonce={ cspNonce }
-                innerHTML={ `
-            * {
-                box-sizing: border-box;
-            }
-        
-            #${ uid } {
-                display: flex;
-                position: fixed;
-                width: 100%;
-                height: 100%;
-                top: 0;
-                left: 0;
-                z-index: 20000;
-                align-items: center;
-                justify-content: center;
-                background-color: rgba(0, 0, 0, 0.4); 
-            }
-            #${ uid } iframe {
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                position: absolute;
-                width: 100%;
-                height: 100%;
-                top: 0;
-                left: 0;
-                transition: opacity .2s ease-in-out;
-            }
-            #qrModal {
-                background: #2F3033;
-                box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.4);
-                border-radius: 16px;                        
-                width: 720px;
-                height: 544px;  
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                flex-direction: column;
-                position: relative;
-            }
-            #close {
-                position: absolute;
-                right: 16px;
-                top: 16px;
-                width: 16px;
-                height: 16px;
-                opacity: 0.6;
-                z-index: 10;
-            }
-            #close:hover {
-                opacity: 1;
-            }
-            #close:before, #close:after {
-                position: absolute;
-                left: 8px;
-                content: ' ';
-                height: 16px;
-                width: 2px;
-                background-color: #FFF;
-            }
-            #close:before {
-                transform: rotate(45deg);
-            }
-            #close:after {
-                transform: rotate(-45deg);
-            }        
-            ` } />
-            <div id="qrModal">
-                <a href="#" id="close" aria-label="close" role="button" onClick={ close } />
-                <node el={ prerenderFrame } />
-                <node el={ frame } />
-            </div>
-                        
-        </div>
-    );
 }
