+++ src/zoid/qr-code/types.js
@@ -1,0 +1,6 @@
+/* @flow */
+
+export type QRCodeProps = {|
+    qrPath : string,
+    cspNonce : ?string
+|};
